Reorganize Demo and add multiple behavior example
