+++ Demo/src/ColoredQuad.hpp
@@ -1,0 +1,11 @@
+#pragma once
+
+#include <BillyEngine.hpp>
+
+class ColoredQuad : public BillyEngine::EntityBehavior {
+   public:
+    SCRIPTABLE_ENTITY(ColoredQuad)
+    BillyEngine::Color color = BillyEngine::Color::white;
+
+    void OnCreate() override;
+};
