+++ Demo/src/MouseChaser.hpp
@@ -1,0 +1,10 @@
+#include <BillyEngine.hpp>
+
+class MouseChaser : public BillyEngine::EntityBehavior {
+   public:
+    SCRIPTABLE_ENTITY(MouseChaser)
+
+    void OnCreate() override;
+
+    void OnUpdate(f32) override;
+};
