+++ Demo/src/PrintName.hpp
@@ -1,0 +1,12 @@
+#pragma once
+
+#include <BillyEngine.hpp>
+
+class PrintName : public BillyEngine::EntityBehavior {
+   public:
+    SCRIPTABLE_ENTITY(PrintName)
+
+    void OnCreate() override {
+        Log.Debug("Heyy I'm new! My name is {}", Tag().Name);
+    }
+};
