+++ Demo/src/Spawner.hpp
@@ -1,0 +1,8 @@
+#include <BillyEngine.hpp>
+
+class Spawner : public BillyEngine::EntityBehavior {
+   public:
+    SCRIPTABLE_ENTITY(Spawner)
+
+    void OnCreate() override;
+};
